# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Building Maintenance API",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:8080/api",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/residents/{id}": {
      "get": {
        "tags": [
          "Resident Management"
        ],
        "summary": "Get a resident by ID",
        "description": "Returns a resident by their unique identifier.",
        "operationId": "getOne_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "UUID of the resident",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Resident found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResidentDto"
                }
              }
            }
          },
          "404": {
            "description": "Resident not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResidentDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Resident Management"
        ],
        "summary": "Update a resident",
        "description": "Updates an existing resident by their ID.",
        "operationId": "update_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "UUID of the resident",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "description": "Resident update data",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResidentUpdateCommandDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Resident updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResidentDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResidentDto"
                }
              }
            }
          },
          "404": {
            "description": "Resident not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResidentDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Resident Management"
        ],
        "summary": "Delete a resident",
        "description": "Deletes a resident by their ID.",
        "operationId": "delete_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "UUID of the resident",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Resident deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Resident not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteResponseDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/residents/{residentId}/requests": {
      "get": {
        "tags": [
          "Resident Maintenance Requests"
        ],
        "summary": "List all maintenance requests for a resident",
        "description": "Returns a list of all maintenance requests for the specified resident.",
        "operationId": "list",
        "parameters": [
          {
            "name": "residentId",
            "in": "path",
            "description": "UUID of the resident",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of maintenance requests",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MaintenanceRequestDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Resident Maintenance Requests"
        ],
        "summary": "Create a new maintenance request for a resident",
        "description": "Creates a new maintenance request for the specified resident.",
        "operationId": "create_2",
        "parameters": [
          {
            "name": "residentId",
            "in": "path",
            "description": "UUID of the resident",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "description": "Maintenance request creation data",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Maintenance request created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaintenanceRequestDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MaintenanceRequestDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/residents/{residentId}/requests/{requestId}": {
      "get": {
        "tags": [
          "Resident Maintenance Requests"
        ],
        "summary": "Get a maintenance request by ID for a resident",
        "description": "Returns a maintenance request by its unique identifier for the specified resident.",
        "operationId": "get",
        "parameters": [
          {
            "name": "residentId",
            "in": "path",
            "description": "UUID of the resident",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "requestId",
            "in": "path",
            "description": "UUID of the maintenance request",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Maintenance request found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaintenanceRequestDto"
                }
              }
            }
          },
          "404": {
            "description": "Maintenance request not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MaintenanceRequestDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Resident Maintenance Requests"
        ],
        "summary": "Cancel a maintenance request for a resident",
        "description": "Cancels a maintenance request for the specified resident.",
        "operationId": "cancel",
        "parameters": [
          {
            "name": "residentId",
            "in": "path",
            "description": "UUID of the resident",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "requestId",
            "in": "path",
            "description": "UUID of the maintenance request",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Maintenance request cancelled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CancelResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Maintenance request not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CancelResponseDto"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Resident Maintenance Requests"
        ],
        "summary": "Update a maintenance request for a resident",
        "description": "Updates an existing maintenance request for the specified resident.",
        "operationId": "update_4",
        "parameters": [
          {
            "name": "residentId",
            "in": "path",
            "description": "UUID of the resident",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "requestId",
            "in": "path",
            "description": "UUID of the maintenance request",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "description": "Maintenance request update data",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Maintenance request updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaintenanceRequestDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MaintenanceRequestDto"
                }
              }
            }
          },
          "404": {
            "description": "Maintenance request not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MaintenanceRequestDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ResidentUpdateCommandDto": {
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string",
            "minLength": 1
          },
          "email": {
            "type": "string",
            "minLength": 1
          },
          "password": {
            "type": "string"
          },
          "unitNumber": {
            "type": "string",
            "minLength": 1
          },
          "unitBlock": {
            "type": "string",
            "minLength": 1
          },
          "residentialComplexId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "email",
          "fullName",
          "unitBlock",
          "unitNumber"
        ]
      },
      "ResidentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "fullName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "unitNumber": {
            "type": "string"
          },
          "unitBlock": {
            "type": "string"
          },
          "residentialComplex": {
            "$ref": "#/components/schemas/ResidentialComplexDto"
          }
        }
      },
      "ResidentialComplexDto": {
        "type": "object",
        "description": "Summary information about a residential complex.",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier of the residential complex."
          },
          "name": {
            "type": "string",
            "description": "Name of the residential complex."
          },
          "address": {
            "type": "string",
            "description": "Address of the residential complex."
          },
          "city": {
            "type": "string",
            "description": "City where the residential complex is located."
          },
          "postalCode": {
            "type": "string",
            "description": "Postal code of the residential complex."
          }
        }
      },
...
      "UpdateRequestDto": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string",
            "minLength": 1
          },
          "scheduledAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "description",
          "scheduledAt"
        ]
      },
      "MaintenanceRequestDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "resident": {
            "$ref": "#/components/schemas/ResidentSummaryDto"
          },
          "description": {
            "type": "string"
          },
          "specialization": {
            "type": "string",
            "enum": [
              "ELECTRICAL",
              "PLUMBING",
              "HVAC",
              "CARPENTRY",
              "OTHER"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "SCHEDULED",
              "IN_PROGRESS",
              "COMPLETED",
              "CANCELLED"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "scheduledAt": {
            "type": "string",
            "format": "date-time"
          },
          "completedAt": {
            "type": "string",
            "format": "date-time"
          },
          "technician": {
            "$ref": "#/components/schemas/TechnicianSummaryDto"
          }
        }
      },
      "ResidentSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "fullName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "unitNumber": {
            "type": "string"
          },
          "unitBlock": {
            "type": "string"
          },
          "residentialComplex": {
            "$ref": "#/components/schemas/ResidentialComplexDto"
          }
        }
      },
      "TechnicianSummaryDto": {
        "type": "object",
        "description": "Summary information about a technician.",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Unique identifier of the technician."
          },
          "fullName": {
            "type": "string",
            "description": "Full name of the technician."
          },
          "email": {
            "type": "string",
            "description": "Email address of the technician."
          }
        }
      },
...
      "CreateRequestDto": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string",
            "minLength": 1
          },
          "specialization": {
            "type": "string",
            "enum": [
              "ELECTRICAL",
              "PLUMBING",
              "HVAC",
              "CARPENTRY",
              "OTHER"
            ]
          },
          "scheduledAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "description",
          "scheduledAt",
          "specialization"
        ]
      },
...
      "DeleteResponseDto": {
        "type": "object",
        "description": "DTO returned after deleting a resource, containing a message about the operation result.",
        "properties": {
          "message": {
            "type": "string",
            "description": "Message describing the result of the delete operation.",
            "example": "Residential complex deleted successfully."
          }
        }
      },
      "CancelResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "JWT Authorization header using the Bearer scheme. Example: 'Bearer {token}'",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}